+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
 ### Changed
 
 - Upgrade phpunit/php-code-coverage from 14.3.0 to 14.3.1
+
+### Added
+
+- Optional product-type filtering through `Client::products()->list($type)`
 
 ## [1.0.0] - 2026-08-13